requirements.txt
python-dotenv
feedparser
requests
schedule
openai